moved the bots folder into pages
