+++ frontend/app/src/pages/bots/Body.tsx
@@ -4,15 +4,15 @@ import { EuiLoadingSpinner } from '@elastic/eui';
 import MaterialIcon from '@material/react-material-icon';
 import { EuiGlobalToastList } from '@elastic/eui';
 import { observer } from 'mobx-react-lite';
-import NoneSpace from '../people/utils/NoneSpace';
-import { Button, Modal, SearchTextInput, Divider } from '../components/common';
-import { useStores } from '../store';
-import { useFuse, useScroll } from '../hooks';
-import { colors } from '../config/colors';
-import FadeLeft from '../components/animated/FadeLeft';
-import { useIsMobile } from '../hooks';
-import Form from '../components/form/bounty';
-import { botSchema } from '../components/form/schema';
+import NoneSpace from '../../people/utils/NoneSpace';
+import { Button, Modal, SearchTextInput, Divider } from '../../components/common';
+import { useStores } from '../../store';
+import { useFuse, useScroll } from '../../hooks';
+import { colors } from '../../config/colors';
+import FadeLeft from '../../components/animated/FadeLeft';
+import { useIsMobile } from '../../hooks';
+import Form from '../../components/form/bounty';
+import { botSchema } from '../../components/form/schema';
 import Bot from './Bot';
 import BotView from './BotView';
 import BotSecret from './utils/BotSecret';
