+++ frontend/app/src/pages/bots/BotView.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import styled from 'styled-components';
 import { useHistory } from 'react-router';
 import { observer } from 'mobx-react-lite';
-import { useStores } from '../store';
+import { useStores } from '../../store';
 
-import { Button, Divider, IconButton } from '../components/common';
-import { useIsMobile } from '../hooks';
+import { Button, Divider, IconButton } from '../../components/common';
+import { useIsMobile } from '../../hooks';
 import Bot from './Bot';
 import BotBar from './utils/BotBar';
 import { BotViewProps } from './interfaces';
