+++ frontend/app/src/pages/bots/interfaces.ts
