+++ frontend/app/src/pages/index.tsx
@@ -4,9 +4,9 @@ import '@material/react-material-icon/dist/material-icon.css';
 import { AppMode } from 'config';
 import { Route, Switch } from 'react-router-dom';
 import { observer } from 'mobx-react-lite';
-import BotsBody from '../bots/Body';
 import PeopleHeader from '../people/main/Header';
 import TokenRefresh from '../people/utils/TokenRefresh';
+import BotsBody from './bots/Body';
 import Body from './tribes/Body';
 import Header from './tribes/Header';
 import { MainLayout } from './MainLayout';
